Rename nav_task and eqa_task usage to nav and eqa respectively
