+++ habitat/datasets/eqa/mp3d_eqa_dataset.py
@@ -12,8 +12,8 @@ from typing import List, Optional
 from habitat.config import Config
 from habitat.core.dataset import Dataset
 from habitat.core.registry import registry
-from habitat.tasks.eqa.eqa_task import EQAEpisode, QuestionData
-from habitat.tasks.nav.nav_task import ObjectGoal, ShortestPathPoint
+from habitat.tasks.eqa.eqa import EQAEpisode, QuestionData
+from habitat.tasks.nav.nav import ObjectGoal, ShortestPathPoint
 
 EQA_MP3D_V1_VAL_EPISODE_COUNT = 1950
 DEFAULT_SCENE_PATH_PREFIX = "data/scene_datasets/"
