+++ test/test_habitat_env.py
@@ -15,7 +15,7 @@ import habitat
 from habitat.config.default import get_config
 from habitat.core.simulator import AgentState
 from habitat.datasets.pointnav.pointnav_dataset import PointNavDatasetV1
-from habitat.tasks.nav.nav_task import (
+from habitat.tasks.nav.nav import (
     NavigationEpisode,
     NavigationGoal,
     StopAction,
